Add a short delay before running Cypress tests
The initial run failed with this error in the first test suite:

```
  1) Checkbox component
       props and methods section
         "before all" hook for "is present":
     CypressError: `cy.visit()` failed trying to load:

http://localhost:6060/

We attempted to make an http request to this URL but the request failed without a response.

We received this error at the network level:

  > Error: ESOCKETTIMEDOUT

Common situations why this would fail:
  - you don't have internet access
  - you forgot to run / boot your web server
  - your web server isn't accessible
  - you have weird network configuration settings on your computer
```

Subsequent tests ran fine, so I'm guessing the style guide server took
a few seconds to start up. Hopefully this'll fix it!